// models/converter-form.model.ts
export type ConverterForm = {
	controller: {
		marginTop: {
			id: string;
			label: string;
			type: string;
			value: number;
		};
		marginBottom: {
			id: string;
			label: string;
			type: string;
			value: number;
		};
		marginLeft: {
			id: string;
			label: string;
			type: string;
			value: number;
		};
		marginRight: {
			id: string;
			label: string;
			type: string;
			value: number;
		};
	};
};